requirements.txt
numpy
Pillow
matplotlib
colorsys
scikit-fuzzy
